Rename download class to IceSat2Data for consistency.
